+++ src/Service/Provider/CidaasProvider.php
@@ -39,7 +39,6 @@ final class CidaasProvider extends ClientProviderContract
                 'scopes',
                 // TODO remove in v6
                 'redirectUri',
-                'roleAssignment',
             ])->setRequired([
                 'organizationUrl',
                 'clientId',