+++ src/Service/Provider/GoogleCloudProvider.php
@@ -37,7 +37,6 @@ final class GoogleCloudProvider extends ClientProviderContract
                 'scopes',
                 // TODO remove in v6
                 'redirectUri',
-                'roleAssignment',
             ])->setRequired([
                 'clientId',
                 'clientSecret',