+++ src/Service/Provider/JiraProvider.php
@@ -36,7 +36,6 @@ final class JiraProvider extends ClientProviderContract
                 'scopes',
                 // TODO remove in v6
                 'redirectUri',
-                'roleAssignment',
             ])->setRequired([
                 'clientId',
                 'clientSecret',