+++ src/Service/Provider/KeycloakProvider.php
@@ -36,7 +36,6 @@ final class KeycloakProvider extends ClientProviderContract
                 'scopes',
                 // TODO remove in v6
                 'redirectUri',
-                'roleAssignment',
             ])->setRequired([
                 'keycloakOidcJson',
             ])->setDefaults([